feat: Comprehensive SEO Overhaul
This commit implements a wide range of SEO improvements to the wedding website, based on the user's request to 'overly optimize' it.

The key changes include:
- Refactored the root layout and homepage from client-side to server-side components to improve crawlability by search engines.
- Introduced a centralized metadata system using Next.js's `metadata` API, including a `metadata.ts` file with SEO-friendly titles, descriptions, and keywords.
- Added a dynamically generated `sitemap.xml` via a `sitemap.ts` file to help search engines discover all pages.
- Created a `robots.txt` file to guide search engine crawlers.
- Integrated JSON-LD structured data for the wedding event into the `metadata` API.
- Optimized on-page content for local SEO, targeting 'Rochester MN weddings' and other related keywords.
- Added new sections for 'About the Venue' and 'Things to Do in Rochester' with rich, descriptive content.
- Ensured all images have descriptive alt tags.
