+++ public/robots.txt
@@ -1,0 +1,16 @@
+# See https://www.robotstxt.org/robotstxt.html for documentation on how to use the robots.txt file
+#
+# Allow all crawlers to access the site.
+User-agent: *
+Disallow: /admin/
+Disallow: /api/
+
+# Allow all content for specific search engines
+User-agent: Googlebot
+Allow: /
+
+User-agent: Bingbot
+Allow: /
+
+# Sitemap location
+Sitemap: https://abbifred.com/sitemap.xml
